backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
pyjwt==2.10.1
python-keycloak==3.9.1
google-auth==2.27.0

# AI / LLM
google-generativeai>=0.8.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6

# Utilities
email-validator==2.1.0
